pyproject.toml
[tool.poetry]
name = "vec2pg"
version = "2024.06.0"
description = "vec2pg"
authors = []
packages = [{ include = "vec2pg", from = "src" }]

[tool.poetry.dependencies]
python = "^3.9"
typer-slim = "^0.12.3"
pinecone-client = "^4.1.0"
psycopg = "^3.1.19"
parse = "^1.20.2"
pgvector = "^0.2.5"
numpy = "^2.0.0"
qdrant-client = "^1.10.1"

[tool.poetry.dev-dependencies]
pre-commit = "^3.5.0"
black = "^24.4"
pytest = "^8.2.2"
flake8 = "^5.0.4"
isort = "^5.10.1"
pytest-cov = "^5.0.0"
commitizen = "^3.22.0"
python-semantic-release = "^9.4.2"
python-dotenv = "^1.0.1"

[tool.poetry.scripts]
vec2pg = "vec2pg.cli:app"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
